Fix the PendingAddOp is not recycled when LedgerHandler closed (#3321)
---

**Motivation**

When adding an entry into a bookie, the entry data lifecycle
is handled by the bookie client. The data buffer will be
released after receiving a corresponding response from the
bookie server. So the user doesn't care about the entry
buffer releasing.
But when the ledgerHandler is closed, the PendingAddOp is not
recycled which leads to the data buffer never being released.
We should release that after the callback executed.

(cherry picked from commit 3d01e6e13cf5410e913859c5c92c002e986df518)